+++ changelog.md
@@ -4,6 +4,7 @@ Script Changes
 - Added considerable functionality to use multiple estimators and multiple
   values.  Useful for comparing two different measurements of the same
   estimator.
+- Updated pimcave.py to take a single estimator
 
 ## 2023-11-15
 - Updated merge.py to use less memory for cumulative estimators